Replace cosmic-text example with Parley
