+++ .gitignore
@@ -1,7 +1,7 @@
 /target
 /scripts/gentest/target
 /benches/target
-/examples/cosmic_text/target
+/examples/parley/target
 /yoga_test_fixtures
 /yoga_test_fixtures_grouped
 /test_fixtures/_scratch
