+++ benches/Cargo.toml
@@ -21,7 +21,7 @@ taffy_03 = { package = "taffy", version = "0.3", optional = true }
 yoga = { version = "0.5.0", optional = true }
 ordered-float = { version = "4", optional = true }
 slotmap = { version = "1.0.6", optional = true }
-parley = "0.7"
+parley = "0.11.1"
 
 [features]
 yoga = ["dep:yoga", "dep:slotmap", "dep:ordered-float"]
