+++ examples/parley/src/image.rs
